refactor(#3196): de-inline standalone dynamic-HOF arms onto a stepper module
Slice 1 of #3196. Lift the standalone dynamic higher-order-function
method-arm switch verbatim out of src/codegen/array-prototype-borrow.ts
into a new sibling module src/codegen/array-like-hof-arms.ts
(emitArrayLikeHofArm, ~766 LOC).

array-prototype-borrow.ts shrinks 1840 -> 1186 LOC (-676). The arm bodies
move verbatim behind a single emitArrayLikeHofArm(ctx, fctx, methodName,
callExpr, arm) entry point; per-call loop-local state (receiver/len/idx/
elem temps, closure info, stepper callbacks) threads through an
ArrayLikeHofArmCtx struct param. borrow.ts keeps only the two call sites.

REFACTOR, zero behavior change. tsc --noEmit clean; check:loc-budget green;
targeted standalone array-HOF vitest pass.

Original extraction by dev-m; recovered, validated, and landed by dev-k
after a box-recovery from swap-death.

✓
